+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Changelog
 ### Added
 - Model validator for validation sub-models
 
+### Changed
+- ModelHelper now supports `yii\base\Model`
+
 ## 1.0.0-rc.4.1 - 2017-12-12
 ### Fixed
 - Action returning a false positive on checks