+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 Changelog
 =========
-## Unreleased
+## 1.0.0-rc.5 - 2017-12-13
 ### Added
 - Model validator for validation sub-models